Restructure backend routes add timetable routes (#1042)
* created routes for user timetables

* fixed returns

* separarted timetable and course routes into the new structure

* removed redundant imports

* fix redundant assert

---------
