+++ server/src/timetable/types.ts
@@ -1,0 +1,16 @@
+export class AddCourseDto {
+  term: string;
+  colour: string;
+}
+export class CourseDetails {
+  id: string;
+  selectedClasses: string[];
+}
+
+export class UserTimetable {
+  id: string;
+  name: string;
+  year: number;
+  term: string;
+  primary: boolean;
+}
