+++ server/src/user/types.ts
@@ -15,12 +15,3 @@ export class UserSettings {
   unscheduleClassesByDefault: boolean;
   hideExamClasses: boolean;
 }
-
-export class AddCourseDto {
-  term: string;
-  colour: string;
-}
-export class CourseDetails {
-  id: string;
-  selectedClasses: string[];
-}
